Added Documentation and cleaned up.
